update build and test

+++ Makefile
@@ -1,17 +1,21 @@
 # cannot use relative path in GOROOT, otherwise 6g not found. For example,
 #   export GOROOT=../go  (=> 6g not found)
 # it is also not allowed to use relative path in GOPATH
-export GOROOT=$(realpath ../go)
-export GOPATH=$(realpath .)
-export PATH := $(GOROOT)/bin:$(GOPATH)/bin:$(PATH)
+ifndef TRAVIS
+	export GOROOT=$(realpath ../go)
+	export GOPATH=$(realpath .)
+	export PATH := $(GOROOT)/bin:$(GOPATH)/bin:$(PATH)
+endif
 
 export PKGNAME=mypkg
 export PKGDIR=${GOPATH}/src/github.com/siongui/${PKGNAME}
 
+ALL_GO_SOURCES=$(shell /bin/sh -c "find *.go | grep -v _test.go")
+
 test: fmt
-	rm -rf ${PKGDIR}
+	if [ -d ${PKGDIR} ]; then rm -rf ${PKGDIR}; fi
 	mkdir -p ${PKGDIR}
-	go test -v buildpkg.go buildpkg_test.go -args -pkgdir=${PKGDIR} -pkgname=${PKGNAME}
+	go test -v ${ALL_GO_SOURCES} buildpkg_test.go -args -pkgdir=${PKGDIR} -pkgname=${PKGNAME}
 	go test -v import_test.go
 
 fmt:
